requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
google-generativeai==0.3.1
jinja2==3.0.1
python-multipart==0.0.5
mangum==0.17.0
aiofiles==0.8.0